fix: get contract size back below limit
To get the contract size below the limit, I had to:
1. increase the number of optimizer runs to max
2. change the solidity version
3. remove several functions

I've also updated the tests where they were broken from the migration to ethers v6, along with removing tests for the deleted solidity functions.

There is a weird bug that prints "[Warning] Invalid Fragment" in the console, although it shouldn't impact the coverage report. It seems to be occur on line 18 of OZGovernorModule.test.ts, but I have no idea why.